+++ src/helper.ts
@@ -1,6 +1,6 @@
 import { useApi, identityApiRef } from '@backstage/core-plugin-api';
 
-export const getRepoName = (e: any) => {
+export const getRepoName = (e: any) : string => {
   if ('github.com/project-slug' in e.entity.metadata.annotations) {
     return e.entity.metadata.annotations['github.com/project-slug'].split('/')[1]
   } else {