+++ src/index.ts
@@ -1,7 +1,5 @@
 // This is our entry point
 export {
-  EntityDORAChangeLeadTimeCard,
-  EntityDORADeploymentFrequency,
-  EntityDORAChangeFailureRate,
-  EntityDORARecoverTime,
+  EntityDORAScoreBoard,
+  EntityDORACharts
 } from './plugin';